migration: Add helper to migrate tarballs to s3
Release tarballs are currently stored in the ompi-www
git repo, which has resulted in a giant web tree that no
one can check out without filling some disk or another.
We're migrating all tarballs to S3, but need to handle
file metadata synchronization.  Using the same json
blobs as the nightly build script simplifies things, so
we need a way to generate said blobs.

This script will build a staged tree of the tarballs and
source RPMS from the existing web tree (or an rsync of
the HostGator tree if you want nightlies as well) suitable
for a recursive s3 cp into the ompi-aws-prodction bucket.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>