jenkins: Add correct logic for updating path
JENKINS_HOME is the master path and has no meaning on the
slaves.  Use either JENKINS_AGENT_HOME (set by an explicit
environment variable in the worker config) or HOME (when
JENKINS_AGENT_HOME is not set) as the guess at the
Remote root directory configuration option (which is where
the autotools will be installed).

Signed-off-by: Brian Barrett <bbarrett@amazon.com>