migration: Sync Windows and MacOS builds
The first version of the script to sync all the tarballs from
the web tree into S3 only handled rpms and tarballs, but there
are also Windows and MacOS builds in the tree as well.  This
patch updates the script to properly add those files into the
S3 list and into the build info files.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>