jenkins: Add Open MPI tarball creation pipeline
Jenkins pipeline for building an Open MPI tarball / release.  Takes
two parameters: a hash / tag to build and an option for whether the
build is a release, pre-release, or scratch.  Release and
pre-release  must have tags as a ref and will push to the release
tree in S3.  Release must not have a greek portion of the version
(a1, b1, rc1, etc.), while pre-release msut have a greek portion
of the version.  The two builds are otherwise identical.  Scratch
builds do not require the ref to be a tag (any git reference will
do), there are no checks that the version number is sane or on
the right branch or anything like that, and the artifacts are pushed
to the open-mpi-scratch S3 bucket, which auto-deletes artifacts after
30 days.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>