wait-for-pr-ci-completion.py: wait for Github PR CI
As the script name implies, wait for a Github PR's CI to complete.
This script does no notification; it just waits -- when the CI
completes, this script completes.  You'll typically want to execute
another command after this script completes, for example:

```
$ ./wait-for-pr-ci-completion.py \
    --pr https://github.com/open-mpi/ompi/pull/5731; \
    pushover CI for PR5731 is done
```

where `pushover` is a script I use to send push notifications to my
phone.

See the comments at the beginning of this script to see its
requirements and how to use it.

This script may get copied out of the Open MPI script repo, so I took
the liberty of including the license in the file.

Signed-off-by: Jeff Squyres <jeff@squyres.com>