dist: Publish SHA256 hashes for tarballs
Update the script that is used by Open MPI (through the tarball
Jenkins job) and hwloc (manually) to publish sha256 hashes in
addition to the md5 and sha1 hashes.  The Open MPI web site has
already been updated to publish the hash if it is available and
has been publishing sha256 hashes for nightlies for a couple of
months.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>