jenkins: Use OS name in autotools cache
We cache Autotools builds in S3 for use by the various Jenkins
builders, to speed up getting the right Autotools for a particular
build.  For some reason, we did not include the OS/Distro name and
version in the cache string.  This was leading to errors due to
missing perl versions and the like.  Add the OS/Distro name and
version to avoid those corner cases.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>