jenkins: Fix autotools situation
Remove the unused agent bring up script that built a static version
of the autotools.

Update the autotools builder that works with Jenkins jobs to
build outside of the workspace, so that builds can persist
across jobs.  This also makes saving builds easier, as the
pathnames to the build don't include a jobname (which can change,
like between preprod and prod versions).

Signed-off-by: Brian Barrett <bbarrett@amazon.com>