jenkins: Add commit email/signed-off-by check jobs
We are trying to migrate our commit email and signed off by
checkers from PHP invoked directly from the GitHub hooks to
Jenkins jobs, so that we have retries and polling and all
that goodness.  This patch contains two pieces of that:

A Groovy script to implement the execution part of the pipeline
job.  This script relies on two parameters from the Jenkins
job configuration for the underlying script to run and the
"checker" name to use on GitHub.

Changes to the previously written Python-based checks, to make
them work under Python2 and to fix some poor design choices
around arguments and output that I had made when telling Jeff
the API for the first implementation.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>