jenkins: update AMI processing scripts
Three major changes to the AMI building scripts:

1) We now use Packer's AMI filter functionality to choose the
latest version of the base AMIs.  This saves lots of time
both in making new AMIs (fewer updates) and in human time
of sometimes having to update the base AMI list.

2) Add ARM AMIs for most operating systems we support.  This
does not include FreeBSD, because there are no ARM64 base
AMIs for FreeBSD11.

3) Increase the root volume size to 16 GiB to try and avoid
running out of disk space so often.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>