Builder.py: fix Coverity submission Python error
dict.keys() returns a dict_keys object, which is not subscriptable.
Hence, if "foo" is a dict, foo.keys()[0] is a Python error.  This led
to this error occurring in the OMPI nightly builds:

```
ERROR: Coverity submission failed: 'dict_keys' object is not subscriptable
```

Instead, convert the returned dict_keys to an iterator and grab the
first item via next().

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>