Coverity: ensure we use the same filename
Use a variable to use the name of the local Coverity tools file.

Sidenote: I discovered that the pre-cached Coverity file on our build
VM was "coverity_tools.tgz", but the script was looking for
"coverity_tool.tgz".  We might as well use the one that was already
cached, so update the filename here in the script to
"coverity_tools.tgz".

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>