openmpi-nightly-tarball: Change Coverity project name
Update it from "Open+MPI" to "Open MPI".  I don't know if a change
occurred on the Coverity server side or if the change to Python
"requests" makes a difference, but using the project name "Open+MPI"
in the requests data causes authentication to fail, but using "Open
MPI" succeeds.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>