Coverity: use wget to download the Coveirty tool
As of 9 Aug 2021, the Coverity tools file is 2+GB.  Downloading it all
into a Python script and then writing it out to disk is not a good
idea on our limited resources AWS VM (meaning: it brings the VM to a
crawl).  From
https://stackoverflow.com/questions/38969164/coverity-scan-for-projects-outside-github
we can use a command line tool to download, instead.  It's not very
Pythonic, but it doesn't bring our VM to its knees.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>