jenkins: make autotools script more general
Refactor the script used by the Open MPI dist tarball builder to
build the Autotools necessary for that tarball to be more general.
The goal is to have a script usable not just for the dist job, but
also for nightly tarballs and CI tests.

The nightly tarballs requirement complicates the autotools script
considerably, as the HWLOC nightly tarballs require an autotools
install, but does not update the required Autotools versions in
the dist script, so we allow setting component versions explicitly
on the command line.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>