nightly: Support relative submodule paths
GitPython's internal submodules support is not a direct call to
the underlying git submodule interface, but does a bunch of
magic that we do not need.  It also does not support relative
paths, which are now used in OMPI's submodule definition.  Switch
to using the underlying git submodule calls directly to fetch
the submodules.

Using the underlying submodules command required us to have a HEAD
that is in refs/ as a local branch.  Change the checkout command
so that the branch we are building is a local branch to stop
irritating git submodule commands.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>