Allow building Sphinx docs
If docs/requirements.txt exists in the source tree, that means that we
have Sphinx documentation, not Pandoc documentation, and we don't need
to do the Pandoc-specific things to make HTML man pages.

Also add some comments about the pending switch on the v5.0.x branch
from Pandoc to Sphinx.  When that happens, we can delete
Pandoc-specific stuff from these scripts.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>