Add pipeline to download coverity tools
The coverity tools are large (1+ GiB) and change infrequently.  Create
a pipeline intended to run weekly to move the tools from Coverity into
S3, so that we don't pay ingress charges every time we run a job.
Unfortunately, I can't find a way to authenticate to the download
server to just call HEAD and get the download time (which would be
the smarter option), so this will have to do.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>