jenkins: Update AMI builders
Jenkins as of Nov 2024 requires at least Java 17 to build, which
is newer than the default for a number of our test OSes.  This
commit gets Java 17 or later installed on all the OSes currently
tested by the Open MPI testes.

Rename jenkins-amis.hcl to jenkins-amis.pkr.hcl and update to HCL2
format, which is easier to read and what packer really wants.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>