jenkins: Add tags to build AMIs
Add some debugging ability to building the Jenkins AMIs.  First, add
a `-d` flag to build_amis to enable Packer's debugging mode.  Second,
tag all built AMIs with two flags: one to indicate that the AMI was
built for jenkins (to make cleanup workflows easier later) and one to
indicate whether the AMI is intended for Production use.  Only if
the `-p` flag is specified will Jenkins automatically pick the AMI
for use.

Signed-off-by: Brian Barrett <bbarrett@amazon.com>